requirements.txt
django
psycopg2-binary
praw
django-extensions
finnhub-python
pdbp
nltk
pandas
lxml
vaderSentiment
djangorestframework